# swagger/openapi.yaml
openapi: 3.0.0
info:
  title: FacePay API
  version: 1.0.0
  description: |
    ## FacePay - Biometric Payment Platform API
    
    FacePay combines advanced biometric authentication with secure payment processing to create a seamless, secure payment experience. Our API provides comprehensive endpoints for user management, payment processing, biometric authentication, and transaction handling.
    
    ### Key Features
    - 🔐 **Biometric Authentication**: WebAuthn-based face recognition and fingerprint authentication
    - 💳 **Multi-Payment Support**: Stripe, crypto, and traditional payment methods
    - 🛡️ **Enhanced Security**: Advanced threat detection and anomaly analysis
    - 📊 **Analytics Dashboard**: Comprehensive payment and security analytics
    - 🔄 **Real-time Processing**: Instant transaction processing and validation
    
    ### Authentication
    Most endpoints require authentication using JWT tokens. Include your token in the Authorization header:
    ```
    Authorization: Bearer YOUR_JWT_TOKEN
    ```
    
    ### Rate Limiting
    - **Standard endpoints**: 1000 requests per hour per user
    - **Authentication endpoints**: 10 requests per minute per IP
    - **Payment endpoints**: 100 requests per hour per user
    
    ### Error Handling
    All endpoints return standardized error responses with appropriate HTTP status codes and detailed error information.
    
  termsOfService: https://facepay.com/terms
  contact:
    name: FacePay API Support
...
servers:
...
paths:
...
  /users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Retrieve the authenticated user's profile information
      operationId: getUserProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
              examples:
                user_profile:
                  summary: Complete user profile
                  value:
                    success: true
                    data:
                      id: clp1abc123def456ghi789
                      email: john.doe@example.com
                      name: John Doe
                      createdAt: "2024-01-01T12:00:00Z"
                      updatedAt: "2024-01-01T12:00:00Z"
                      biometricData:
                        - id: bio_abc123
                          type: face
                          isActive: true
                          createdAt: "2024-01-01T12:00:00Z"
                      paymentMethods:
                        - id: pm_abc123
                          type: card
                          provider: stripe
                          isDefault: true
                          createdAt: "2024-01-01T12:00:00Z"
                      _count:
                        transactions: 42
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Users
      summary: Update user profile
      description: Update the authenticated user's profile information
      operationId: updateUserProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: John Updated Doe
                email:
                  type: string
                  format: email
                  example: john.updated@example.com
            examples:
              profile_update:
                summary: Update profile
                value:
                  name: John Updated Doe
                  email: john.updated@example.com
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Email already taken
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Users
      summary: Delete user account
      description: Permanently delete the authenticated user's account and all associated data
      operationId: deleteUserAccount
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: null
                      message:
                        type: string
                        example: Account deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Payment Endpoints
...
  /transactions/history:
    get:
      tags:
        - Transactions
      summary: Get detailed transaction history
      description: Get comprehensive transaction history with advanced filtering
      operationId: getTransactionHistory
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: amountMin
          in: query
          schema:
            type: number
        - name: amountMax
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Transaction history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from the login endpoint. Include it in the Authorization header:
        ```
        Authorization: Bearer YOUR_JWT_TOKEN
        ```
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for external integrations and webhooks
    
    BiometricAuth:
      type: oauth2
      description: Biometric authentication using WebAuthn
      flows:
        authorizationCode:
          authorizationUrl: /api/webauthn/authenticate/start
          tokenUrl: /api/webauthn/authenticate/complete
          scopes:
            biometric:read: Read biometric authentication data
            biometric:write: Register and manage biometric credentials
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          additionalProperties: true
        message:
          type: string
          example: Operation completed successfully
      required:
        - success
        - data
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: An error occurred
        code:
          type: string
          example: INTERNAL_ERROR
        details:
          type: object
          additionalProperties: true
      required:
        - success
        - error
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Validation failed
        code:
          type: string
          example: VALIDATION_ERROR
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Email is required

    # User Schemas
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
          example: clp1abc123def456ghi789
        email:
          type: string
          format: email
          description: User email address
          example: john.doe@example.com
        name:
          type: string
          description: User full name
          example: John Doe
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: 2024-01-01T12:00:00Z
        updatedAt:
          type: string
          format: date-time
          description: Last profile update timestamp
          example: 2024-01-01T12:00:00Z
        lastLogin:
          type: string
          format: date-time
          description: Last login timestamp
          example: 2024-01-01T12:00:00Z
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            biometricData:
              type: array
              items:
                $ref: '#/components/schemas/BiometricData'
              description: User's registered biometric data
            paymentMethods:
              type: array
              items:
                $ref: '#/components/schemas/PaymentMethod'
              description: User's payment methods
            _count:
              type: object
              properties:
                transactions:
                  type: integer
                  description: Total number of transactions
                  example: 42
...
    PaymentMethod:
      type: object
      properties:
        id:
          type: string
          description: Payment method identifier
          example: pm_abc123def456
        type:
          type: string
          enum: [card, bank_account, crypto, biometric]
          description: Type of payment method
          example: card
        provider:
          type: string
          enum: [stripe, paypal, crypto, internal]
          description: Payment provider
          example: stripe
        isDefault:
          type: boolean
          description: Whether this is the default payment method
          example: true
        details:
          type: object
          additionalProperties: true
          description: Provider-specific payment method details
          example:
            last4: "4242"
            brand: visa
            exp_month: 12
            exp_year: 2025
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          description: Transaction identifier
          example: txn_abc123def456
        userId:
          type: string
          description: User identifier
          example: usr_abc123def456
        amount:
          type: number
          description: Transaction amount
          example: 29.99
        currency:
          type: string
          description: Currency code
          example: USD
        status:
          type: string
          enum: [pending, completed, failed, refunded]
          description: Transaction status
          example: completed
        paymentMethodId:
          type: string
          description: Payment method used
          example: pm_abc123def456
        description:
          type: string
          description: Transaction description
          example: Premium subscription
        metadata:
          type: object
          additionalProperties: true
          description: Additional transaction data
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
...
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        pagination:
          type: object
          properties:
            page:
              type: integer
              description: Current page number
              example: 1
            limit:
              type: integer
              description: Items per page
              example: 20
            total:
              type: integer
              description: Total number of items
              example: 100
            totalPages:
              type: integer
              description: Total number of pages
              example: 5
            hasNext:
              type: boolean
              description: Whether there are more pages
              example: true
            hasPrev:
              type: boolean
              description: Whether there are previous pages
              example: false

    # Biometric Schemas
    BiometricData:
      type: object
      properties:
        id:
          type: string
          description: Biometric data identifier
          example: bio_abc123def456
        type:
          type: string
          enum: [face, fingerprint, voice]
          description: Biometric data type
          example: face
        isActive:
          type: boolean
          description: Whether the biometric data is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: Registration timestamp

    # WebAuthn Schemas
...
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Authentication required
            code: UNAUTHORIZED
    ForbiddenError:
      description: Insufficient permissions to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Insufficient permissions
            code: FORBIDDEN
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            success: false
            error: Validation failed
            code: VALIDATION_ERROR
            details:
              - field: email
                message: Email is required
              - field: password
                message: Password must be at least 8 characters
    InternalServerError:
      description: Internal server error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Internal server error
            code: INTERNAL_ERROR
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination (starts from 1)
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page (maximum 100)
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
security:
  - BearerAuth: []